Refactor file writing process. We now write to an internal buffer and then write that to a file after the command has finished.
