+++ comdirect/cmd/depot.go
@@ -38,7 +38,7 @@ func depot(cmd *cobra.Command, args []string) {
 }
 
 func printDepotsCSV(depots *comdirect.Depots) {
-	table := csv.NewWriter(getOutputFile())
+	table := csv.NewWriter(getOutputBuffer())
 	table.Write([]string{"DEPOT ID", "DISPLAY ID", "HOLDER NAME", "CLIENT ID"})
 	for _, d := range depots.Values {
 		table.Write([]string{d.DepotId, d.DepotDisplayId, d.HolderName, d.ClientId})
@@ -47,7 +47,7 @@ func printDepotsCSV(depots *comdirect.Depots) {
 }
 
 func printDepotsTable(depots *comdirect.Depots) {
-	table := tablewriter.NewWriter(getOutputFile())
+	table := tablewriter.NewWriter(getOutputBuffer())
 	table.SetHeader([]string{"DEPOT ID", "DISPLAY ID", "HOLDER NAME", "CLIENT ID"})
 	table.SetBorders(tablewriter.Border{Left: true, Top: false, Right: true, Bottom: false})
 	table.SetCenterSeparator("|")
