+++ comdirect/cmd/position.go
@@ -39,7 +39,7 @@ func position(cmd *cobra.Command, args []string) {
 }
 
 func printPositionsCSV(positions *comdirect.DepotPositions) {
-	table := csv.NewWriter(getOutputFile())
+	table := csv.NewWriter(getOutputBuffer())
 	table.Write([]string{"POSITION ID", "WKN", "QUANTITY", "CURRENT PRICE", "PREVDAY %", "PURCHASE %", "PURCHASE", "CURRENT"})
 	for _, d := range positions.Values {
 		table.Write([]string{d.PositionId, d.Wkn, d.Quantity.Value, d.CurrentPrice.Price.Value, d.ProfitLossPrevDayRel, d.ProfitLossPurchaseRel, d.PurchaseValue.Value, d.CurrentValue.Value})
@@ -48,7 +48,7 @@ func printPositionsCSV(positions *comdirect.DepotPositions) {
 }
 
 func printPositionsTable(depots *comdirect.DepotPositions) {
-	table := tablewriter.NewWriter(getOutputFile())
+	table := tablewriter.NewWriter(getOutputBuffer())
 	table.SetHeader([]string{"POSITION ID", "WKN", "QUANTITY", "CURRENT PRICE", "PREVDAY %", "PURCHASE %", "PURCHASE", "CURRENT"})
 	table.SetBorders(tablewriter.Border{Left: true, Top: false, Right: true, Bottom: false})
 	table.SetCenterSeparator("|")
