+++ comdirect/cmd/transaction.go
@@ -102,7 +102,7 @@ func getTransactionsSince(since string, client *comdirect.Client, accountID stri
 }
 
 func printTransactionCSV(transactions *comdirect.AccountTransactions) {
-	table := csv.NewWriter(getOutputFile())
+	table := csv.NewWriter(getOutputBuffer())
 	table.Write(transactionHeader)
 	for _, t := range transactions.Values {
 		holderName := t.Remitter.HolderName
@@ -116,7 +116,7 @@ func printTransactionCSV(transactions *comdirect.AccountTransactions) {
 	table.Flush()
 }
 func printTransactionTable(transactions *comdirect.AccountTransactions) {
-	table := tablewriter.NewWriter(getOutputFile())
+	table := tablewriter.NewWriter(getOutputBuffer())
 	table.SetHeader(transactionHeader)
 	table.SetColumnAlignment([]int{tablewriter.ALIGN_LEFT, tablewriter.ALIGN_LEFT, tablewriter.ALIGN_LEFT, tablewriter.ALIGN_LEFT, tablewriter.ALIGN_LEFT, tablewriter.ALIGN_RIGHT})
 	table.SetBorders(tablewriter.Border{Left: true, Top: false, Right: true, Bottom: false})
